Embed structure.
